# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photo directly from your PC, and they will be visible to everyone following you
  version: 1.0.0

#============================================TAGS PART==============================================#
...
security:
  - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Bearer token for the WASAPhoto API.
      scheme: bearer

#============================================SCHEMAS PART==============================================#
  schemas:
...
    Code-Identifier:
      title: Code-random
      type: object
      description: |-
        General random generated string, which will be used as a unique identifier.
        It is a unique code that identifier uniquely:
          - the user
          - the photo
          - the comment
      properties:
        identifier: 
          type: string
          description: unique identifier generatedr randomly
          pattern: '^.*?$'
          minLength: 64
          maxLength: 64
          example: "115981bedf5cc6956330a1e4c394526a8839e8a3ec3adf0fre3eedd59c759a58"
      
    
#_____________________________USER FIELD____________________________________________
#-------------------------------------------------------------------------------- 
...
    LikeID:
      title: Like ID
      description: It's a object that contains an identifier code of like
      type: object
      properties:
        code-like: 
          $ref: "#/components/schemas/Code-Identifier"
        
#-------------------------------------------------------------------------------- 
    Like:
      title: Like
      description: |-
        Like is an object which are put in the photo by Users. It is composed by:
          - Like ID 
          - Owner like
      type: object
      properties:
        like-id:
          $ref: "#/components/schemas/LikeID"
  
        photo-id:
          $ref: "#/components/schemas/ImageID"
        
        from-user: 
          $ref: "#/components/schemas/Username/properties/username-string"
        


#_____________________________PHOTO FIELD____________________________________
#----------------------------------------------------------------------------

#-------------------------------------------------------------------------------- 
...
    ImageID:
      title: ImageID
      description: It's an object that contains an identifier code of image
      type: object
      properties:
        code-image: 
          $ref: "#/components/schemas/Code-Identifier"
          
#-------------------------------------------------------------------------------- 
...
    Error:
      title: Error
      type: object
      description: |-
        An error message that contains:
          - Code 
          - Message
      properties:
        code-error:
          type: integer
          description: The error code
          minimum: 3
          maximum: 3

        message-error:
          type: string
          description: The error message
          minLength: 1
          maxLength: 256
          pattern: ^[^\\]{1,256}$

#============================================PARAMETERS PART==============================================#
  parameters:
    user-id:
      name: user-id
      in: path
      description: The ID of a user to get information about him/her
      required: true
      schema: 
        $ref: "#/components/schemas/Username/properties/username-string"
      
#--------------------------------------------------------------------------------
...
    photo-id:
      name: photo-id
      in: path
      description: The ID photo identifier
      required: true
      schema: 
        $ref: "#/components/schemas/Code-Identifier/properties/identifier"
        
#--------------------------------------------------------------------------------
...
    like-id:
      name: like-id
      in: path
      description: The ID like identifier
      required: true
      schema: 
        $ref: "#/components/schemas/Code-Identifier/properties/identifier"


#============================================PATHS PART==============================================#
paths:
...
  /users/{user-id}/homepage/{photo-id}/likes/{like-id}:
  # Metodi: PUT --> Mettere like DELETE --> togliere il like
    parameters: 
      - $ref: "#/components/parameters/user-id"
      - $ref: "#/components/parameters/photo-id"
      - $ref: "#/components/parameters/like-id"
    # methods
    put:
      tags: ["like", "photo"]
      summary: Put like photo
      description: the user like a photo so the user must authenticated
      operationId: likePhoto
      
      requestBody:
        description: the like to put
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Like"
            
      responses:
        "201":
          description: "The photo has been liked"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Like"

        "400":
          description: Invalid photo/like/user id
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "403":
          description: The like is forbidden because the user was banned
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: The resources are not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


    delete:
      tags: ["stream", "like", "photo"]
      summary: Unlike the Photo
      description: the user unlike a photo so the user must authenticated
      operationId: unlikePhoto
      
      responses:
        "204":
          description: "The photo has been unliked"

        "400":
          description: Invalid photo/like/user id
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


        "403":
          description: The unlike is forbidden
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: The like-id is not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}

#-------------------------------------------------------------------------------- 
...